# docs/swagger.yaml
definitions:
...
  controllers.AnchorWorkCount:
    properties:
      rate:
        type: string
      sign_count:
        type: string
    type: object
...
  controllers.JsonResult:
    properties:
      code:
        type: integer
      data:
        type: object
      err_msg:
        type: string
    required:
    - code
    type: object
...
  dao.RewardConfig:
    properties:
      regulation_cycle:
        description: 调控周期
        type: integer
      sign_reward:
        description: 单笔签名奖励
        type: string
      source_chain_id:
        type: integer
      target_chain_id:
        type: integer
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /anchor/work/count:
    get:
      consumes:
      - application/json
      parameters:
      - description: 锚定节点id
        in: query
        name: anchor_node_id
        required: true
        type: string
      - description: 节点id
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/controllers.AnchorWorkCount'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 本期总签名数及签名工作量占比
      tags:
      - GetAnchorWorkCount
...
  /reward/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 锚定节点id
        in: formData
        name: anchor_node_id
        required: true
        type: integer
      - description: 节点id
        in: formData
        name: node_id
        required: true
        type: integer
      - description: 钱包id
        in: formData
        name: wallet_id
        required: true
        type: integer
      - description: 钱包密码
        in: formData
        name: password
        required: true
        type: string
      - description: 奖励金额
        in: formData
        name: reward
        required: true
        type: string
      - description: 奖励币种
        in: formData
        name: coin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 新增奖励发放
      tags:
      - AddSignReward
  /reward/anchor/single/:
    get:
      consumes:
      - application/json
      parameters:
      - description: 锚定节点id
        in: query
        name: anchor_node_id
        required: true
        type: integer
      - description: 节点id
        in: query
        name: node_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dao.RewardConfig'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取单笔签名奖励(根据节点和锚定节点)
      tags:
      - GetSignRewardByAnchorNode
  /reward/chain:
    get:
      consumes:
      - application/json
      parameters:
      - description: 锚定节点id
        in: query
        name: anchor_node_id
        required: true
        type: string
      - description: 节点id
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 单笔签名奖励
      tags:
      - GetChainReward
  /reward/chain/single/:
    get:
      consumes:
      - application/json
      parameters:
      - description: 发起链id
        in: query
        name: source_chain_id
        required: true
        type: integer
      - description: 目标链id
        in: query
        name: target_chain_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dao.RewardConfig'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取单笔签名奖励（根据发起链和目标链）
      tags:
      - GetSignRewardBySourceAndTarget
...
  /reward/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 锚定节点id
        in: formData
        name: anchor_node_id
        required: true
        type: string
      - description: 当前页
        in: formData
        name: current_page
        required: true
        type: string
      - description: 页的记录数
        in: formData
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 签名奖励
      tags:
      - ListSignReward
...
  /reward/total:
    get:
      consumes:
      - application/json
      parameters:
      - description: 锚定节点id
        in: query
        name: anchor_node_id
        required: true
        type: string
      - description: 节点id
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 剩余奖池总额
      tags:
      - GetTotalReward
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"